pyproject.toml
[project]
name = "audio_app"
version = "0.1.0"
description = "Aplicación Django para gestión de audios y transcripciones con Whisper"
authors = [{ name = "Guillermo Peralta-Godoy" }]
requires-python = ">=3.10"

# Dependencias principales
dependencies = [
    # --- Web y framework principal ---
    "Django==5.1.3",
    "django-allauth==65.8.1",
    "djangorestframework==3.15.2",
    "asgiref==3.8.1",
    "sqlparse==0.5.3",
    "psycopg2-binary==2.9.10",

    # --- Autenticación y HTTP ---
    "requests==2.32.3",
    "urllib3==2.4.0",
    "idna==3.10",
    "certifi==2025.4.26",
    "charset-normalizer==3.4.2",
    "cryptography==45.0.3",
    "cffi==1.17.1",
    "PyJWT==2.10.1",
    "pycparser==2.22",

    # --- Procesamiento de audio (Whisper + PyTorch) ---
    "torch==2.4.1",
    "openai-whisper==20250625",
    "tqdm==4.67.1",
    "numpy==1.24.4",
    "scipy==1.10.1",
    "soundfile==0.12.1",
    "numba==0.58.1",
    "llvmlite==0.41.1",

    # --- Utilidades ---
    "filelock>=3.12.0",
    "importlib-metadata>=8.0",
    "jinja2>=3.1.3",
    "fsspec>=2024.9.0",
    "more-itertools>=10.5.0",
]

# Evita que setuptools intente descubrir paquetes (no es una librería)
[tool.setuptools]
packages = []

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"